refactor: modify parkingFee propertie type

+++ src/main/java/com/pablords/parking/adapters/inbound/http/dtos/CheckoutResponseDTO.java
@@ -14,5 +14,5 @@ public class CheckoutResponseDTO {
     @JsonProperty
     private LocalDateTime checkOutTime;
     @JsonProperty
-    private long parkingFee;
+    private Double parkingFee;
 }
